+++ src/pages/gift/Gift.js
@@ -1,5 +1,5 @@
 import { filter } from 'lodash';
-// import { sentenceCase } from 'change-case';
+// import { sentenceCase } from 'change-case';?
 import { useState } from 'react';
 import { Link as RouterLink } from 'react-router-dom';
 // material
@@ -18,23 +18,22 @@ import {
   TablePagination,
 } from '@mui/material';
 // components
-import Page from '../components/Page';
-// import Label from '../components/Label';
-import Scrollbar from '../components/Scrollbar';
-import Iconify from '../components/Iconify';
-import SearchNotFound from '../components/SearchNotFound';
-import { UserListHead, UserListToolbar, UserMoreMenu } from '../sections/@dashboard/user';
+import Page from '../../components/Page';
+import Scrollbar from '../../components/Scrollbar';
+import Iconify from '../../components/Iconify';
+import SearchNotFound from '../../components/SearchNotFound';
+import { UserListHead, UserListToolbar, UserMoreMenu } from '../../sections/@dashboard/user';
 // mock
-import USERLIST from '../_mock/user';
+import USERLIST from '../../_mock/user';
 
 // ----------------------------------------------------------------------
 
 const TABLE_HEAD = [
   { id: 'no', label: 'No', alignRight: false },
   { id: 'name', label: 'Name', alignRight: false },
-  { id: 'description', label: 'Description', alignRight: false },
+  { id: 'type', label: 'Type', alignRight: false },
+  { id: 'createDate', label: 'Create Date', alignRight: false },
   { id: 'updateDate', label: 'Update Date', alignRight: false },
-  { id: '' },
 ];
 
 // ----------------------------------------------------------------------
@@ -131,14 +130,14 @@ export default function User() {
   const isUserNotFound = filteredUsers.length === 0;
 
   return (
-    <Page title="Condition">
+    <Page title="Gift">
       <Container>
         <Stack direction="row" alignItems="center" justifyContent="space-between" mb={5}>
           <Typography variant="h4" gutterBottom>
-            Condition
+            Gift
           </Typography>
           <Button variant="contained" component={RouterLink} to="#" startIcon={<Iconify icon="eva:plus-fill" />}>
-            New Condition
+            New Gift
           </Button>
         </Stack>
 
@@ -159,19 +158,19 @@ export default function User() {
                 />
                 <TableBody>
                   {filteredUsers.slice(page * rowsPerPage, page * rowsPerPage + rowsPerPage).map((row) => {
-                    const { id, name, description, updateDate, avatarUrl } = row;
-                    // const isItemSelected = selected.indexOf(name) !== -1;
-                    // Bỏ trong return -> <TableRow>
-                    // role="checkbox"
-                    // selected={isItemSelected}
-                    // aria-checked={isItemSelected}
+                    const { id, no, name, type, createDate, updateDate, avatarUrl } = row;
+                    const isItemSelected = selected.indexOf(name) !== -1;
 
-                    // Bỏ trong return -> <TableRow>
-                    // <TableCell padding="checkbox">
-                    //   <Checkbox checked={isItemSelected} onChange={(event) => handleClick(event, name)} />
-                    // </TableCell>;
                     return (
-                      <TableRow hover key={id} tabIndex={-1}>
+                      <TableRow
+                        hover
+                        key={id}
+                        tabIndex={-1}
+                        role="checkbox"
+                        selected={isItemSelected}
+                        aria-checked={isItemSelected}
+                      >
+                        <TableCell align="left">{no}</TableCell>
                         <TableCell component="th" scope="row" padding="none">
                           <Stack direction="row" alignItems="center" spacing={2}>
                             <Avatar alt={name} src={avatarUrl} />
@@ -180,7 +179,8 @@ export default function User() {
                             </Typography>
                           </Stack>
                         </TableCell>
-                        <TableCell align="left">{description}</TableCell>
+                        <TableCell align="left">{type}</TableCell>
+                        <TableCell align="left">{createDate}</TableCell>
                         <TableCell align="left">{updateDate}</TableCell>
                         <TableCell align="right">
                           <UserMoreMenu />
