+++ src/pages/promotion/NewPromotion.js
@@ -1,0 +1,8 @@
+import Page from '../../components/Page';
+import DashboardSidebar from '../../layouts/dashboard/DashboardSidebar';
+
+// ----------------------------------------------------------------------
+
+export default function NewPromotion() {
+  return <Page title="New Promotion">Hello</Page>;
+}
