+++ src/pages/promotion/Promotion.js
@@ -18,25 +18,25 @@ import {
   TablePagination,
 } from '@mui/material';
 // components
-import Page from '../components/Page';
-import Label from '../components/Label';
-import Scrollbar from '../components/Scrollbar';
-import Iconify from '../components/Iconify';
-import SearchNotFound from '../components/SearchNotFound';
-import { UserListHead, UserListToolbar, UserMoreMenu } from '../sections/@dashboard/user';
+import Page from '../../components/Page';
+import Label from '../../components/Label';
+import Scrollbar from '../../components/Scrollbar';
+import Iconify from '../../components/Iconify';
+import SearchNotFound from '../../components/SearchNotFound';
+import { UserListHead, UserListToolbar, UserMoreMenu } from '../../sections/@dashboard/user';
 // mock
-import USERLIST from '../_mock/user';
+import USERLIST from '../../_mock/user';
 
 // ----------------------------------------------------------------------
 
 const TABLE_HEAD = [
   { id: 'no', label: 'No', alignRight: false },
   { id: 'name', label: 'Name', alignRight: false },
-  { id: 'actionName', label: 'Action Name', alignRight: false },
-  { id: 'total', label: 'Total', alignRight: false },
-  { id: 'repemded', label: 'Repemded', alignRight: false },
-  { id: 'used', label: 'Used', alignRight: false },
-  { id: '' },
+  { id: 'type', label: 'Type', alignRight: false },
+  { id: 'action', label: 'Action', alignRight: false },
+  { id: 'start', label: 'Start', alignRight: false },
+  { id: 'isVerified', label: 'Verified', alignRight: false },
+  { id: 'status', label: 'Status', alignRight: false },
 ];
 
 // ----------------------------------------------------------------------
@@ -133,14 +133,19 @@ export default function User() {
   const isUserNotFound = filteredUsers.length === 0;
 
   return (
-    <Page title="Voucher">
+    <Page title="Promotion">
       <Container>
         <Stack direction="row" alignItems="center" justifyContent="space-between" mb={5}>
           <Typography variant="h4" gutterBottom>
-            Voucher
+            Promotion
           </Typography>
-          <Button variant="contained" component={RouterLink} to="#" startIcon={<Iconify icon="eva:plus-fill" />}>
-            New Voucher
+          <Button
+            variant="contained"
+            component={RouterLink}
+            to="/new-promotion"
+            startIcon={<Iconify icon="eva:plus-fill" />}
+          >
+            New Promotion
           </Button>
         </Stack>
 
@@ -161,11 +166,11 @@ export default function User() {
                 />
                 <TableBody>
                   {filteredUsers.slice(page * rowsPerPage, page * rowsPerPage + rowsPerPage).map((row) => {
-                    const { id, name, actionName, total, repemded, avatarUrl, used } = row;
-                    // const isItemSelected = selected.indexOf(name) !== -1;
+                    const { id, name, action, type, start, status, avatarUrl, isVerified } = row;
 
                     return (
-                      <TableRow hover key={id}>
+                      <TableRow hover key={id} tabIndex={-1}>
+                        <TableCell align="left">{id}</TableCell>
                         <TableCell component="th" scope="row" padding="none">
                           <Stack direction="row" alignItems="center" spacing={2}>
                             <Avatar alt={name} src={avatarUrl} />
@@ -174,15 +179,15 @@ export default function User() {
                             </Typography>
                           </Stack>
                         </TableCell>
-                        <TableCell align="left">{actionName}</TableCell>
-                        <TableCell align="left">{total}</TableCell>
-                        <TableCell align="left">{repemded}</TableCell>
+                        <TableCell align="left">{type}</TableCell>
+                        <TableCell align="left">{action}</TableCell>
+                        <TableCell align="left">{start}</TableCell>
+                        <TableCell align="left">{isVerified ? 'Yes' : 'No'}</TableCell>
                         <TableCell align="left">
-                          <Label variant="ghost" color={(used === 'banned' && 'error') || 'success'}>
-                            {sentenceCase(used)}
+                          <Label variant="ghost" color={(status === 'banned' && 'error') || 'success'}>
+                            {sentenceCase(status)}
                           </Label>
                         </TableCell>
-
                         <TableCell align="right">
                           <UserMoreMenu />
                         </TableCell>
